third_party/blink/web_tests/http/tests/origin_trials/webexposed/contacts-manager-extra-properties-trial-interfaces.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 ContactsManagerExtraProperties --expire-timestamp=2000000000
-->
<meta http-equiv="origin-trial" content="Ag08kQBxaZIHRn4szFwJFlEfUhOPCV2CZtU7mqRMQqxt3kGuuPu9Qom6hRdF+vAQIBcHOrDw0d4ajhMzeUETXQ0AAABmeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiQ29udGFjdHNNYW5hZ2VyRXh0cmFQcm9wZXJ0aWVzIiwgImV4cGlyeSI6IDIwMDAwMDAwMDB9" />
<title>Contacts Manager Extra Properties - exposed by origin trial</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/origin-trials-helper.js"></script>
<script>

test(() => {
  OriginTrialsHelper.check_properties_exist(this, { 'ContactsManager': ['select', 'getProperties'] });
}, 'Contact API interfaces and properties in Origin-Trial enabled document.');

promise_test(async () => {
  const properties = await navigator.contacts.getProperties();
  assert_true(properties.includes('address'));
  assert_true(properties.includes('icon'));
}, '`address` and `icon` properties are exposed.');

</script>
